docgen/requirements.txt
google-genai==0.3.0
PyMuPDF==1.24.1
python-docx==0.8.11
pytesseract==0.3.10
Pillow==11.0.0
python-dotenv==1.0.0
reportlab==4.0.9
